Добавлена валидация данных, изменения в сортировке/фильтрации/пагинации. (Request)

+++ TechRequest.API/Controllers/RequestController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using TechRequest.API.Dtos.Request;
 using TechRequest.API.Interfaces;
+using TechRequest.API.Parameters;
 
 namespace TechRequest.API.Controllers
 {
@@ -15,6 +16,9 @@ namespace TechRequest.API.Controllers
         {
             try
             {
+                if(!ModelState.IsValid)
+                    return BadRequest(ModelState);
+
                 var requests = await _requestRepository.GetAllAsync();
                 return Ok(requests);
             }
@@ -24,11 +28,14 @@ namespace TechRequest.API.Controllers
             }
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
             try
             {
+                if (!ModelState.IsValid)
+                    return BadRequest(ModelState);
+
                 var request = await _requestRepository.GetByIdAsync(id);
                 return Ok(request);
             }
@@ -43,6 +50,9 @@ namespace TechRequest.API.Controllers
         {
             try
             {
+                if (!ModelState.IsValid)
+                    return BadRequest(ModelState);
+
                 var request = await _requestRepository.CreateAsync(requestDto);
                 return Ok(request);
             }
@@ -53,32 +63,48 @@ namespace TechRequest.API.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] RequestUpdationDto requestUpdationDto)
         {
-            var requestDto = await _requestRepository.UpdateAsync(id, requestUpdationDto);
-
-            if (requestDto == null)
+            try
             {
-                return NotFound();
-            }
+                if (!ModelState.IsValid)
+                    return BadRequest(ModelState);
 
-            return Ok(requestDto);
+                var requestDto = await _requestRepository.UpdateAsync(id, requestUpdationDto);
+
+                if (requestDto == null)
+                    return NotFound();
+
+                return Ok(requestDto);
+            }
+            catch (Exception ex)
+            { 
+                return BadRequest(ex.Message); 
+            }
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
-            var request = await _requestRepository.DeleteAsync(id);
-
-            if (request == null)
+            try
             {
-                return NotFound();
-            }
+                if (!ModelState.IsValid)
+                    return BadRequest(ModelState);
 
-            //return Ok(request);
-            return NoContent();
+                var request = await _requestRepository.DeleteAsync(id);
+
+                if (request == null)
+                    return NotFound();
+
+                //return Ok(request);
+                return NoContent();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
     }
 }
